Remove deprecated yaml config
In favor of configuration via env vars and sys props.
