+++ dd-trace/src/main/java/com/datadoghq/trace/resolver/DDTracerResolver.java
@@ -29,16 +29,6 @@ public class DDTracerResolver extends TracerResolver {
   protected Tracer resolve() {
     log.info("Creating the Datadog Tracer from the resolver");
 
-    final TracerConfig tracerConfig =
-        FactoryUtils.loadConfigFromFilePropertyOrResource(
-            DDTracerFactory.SYSTEM_PROPERTY_CONFIG_PATH,
-            DDTracerFactory.CONFIG_PATH,
-            TracerConfig.class);
-
-    if (tracerConfig != null) {
-      return DDTracerFactory.createFromConfigurationFile();
-    } else {
-      return new DDTracer();
-    }
+    return new DDTracer();
   }
 }
