+++ dd-trace/src/test/groovy/com/datadoghq/trace/DDTraceConfigTest.groovy
@@ -141,5 +141,7 @@ class DDTraceConfigTest extends Specification {
     tracer.serviceName == "unnamed-java-app"
     tracer.sampler instanceof AllSampler
     tracer.writer.toString() == "DDAgentWriter { api=DDApi { tracesEndpoint=http://localhost:8126/v0.3/traces } }"
+
+    tracer.spanContextDecorators.size() == 2
   }
 }
